chore: close the plan items left open in phases 22, 25 and 26
An audit of the plan against the tree found five tasks partially done:

- CategoryDonut is extracted, taking ChartTooltip with it into its own file
  since the cash-flow chart uses it too. Dashboard.jsx: 754 -> 650 lines.
- ChatMessage moves to schemas/ai.py. It was the only request schema defined
  inside a router, against the layout AGENTS.md declares.
- inputCls existed three times with two different values. The shadcn one is now
  shared as shadcnInputCls; Auth keeps its taller, rounder variant under the
  honest name authInputCls instead of a third thing called inputCls.
- The AGENTS.md rate-limit decision is rewritten. The review was right that
  calling the shared-bucket cost "colateral" understated it: the same proxy IP
  for everyone turns the brute-force defence into a public availability switch.
  Still accepted, now with the real cost and the trigger to revisit written down.

One planned item is deliberately NOT done: removing the jsxInject from
vitest.config.js. The suggestion assumed it was dead weight under React 19, but
Vitest transforms JSX with the classic runtime, so dropping it fails four suites
with "React is not defined". The line stays, with a comment saying why and that
it was verified by removing it. The two unused `import * as React` lines it made
redundant are gone.
