+++ frontend/src/components/ui/dialog.jsx
@@ -1,6 +1,5 @@
 "use client"
 
-import * as React from "react"
 import { Dialog as DialogPrimitive } from "@base-ui/react/dialog"
 
 import { cn } from "@/lib/utils"
