+++ frontend/src/components/ui/input.jsx
@@ -1,4 +1,3 @@
-﻿import * as React from "react"
 import { Input as InputPrimitive } from "@base-ui/react/input"
 
 import { cn } from "@/lib/utils"
